# src/features/documents/types/index.ts
import { z } from 'zod';
import { Party } from "@/features/parties/types";
import { Product } from "@/features/products/types";

export const DocumentTypeEnum = z.enum(['INVOICE', 'CREDIT_NOTE', 'DEBIT_NOTE', 'PAYROLL', 'PAYROLL_ADJUST', 'DOC_SUPPORT', 'RECEIPT', 'QUOTATION', 'SALES_ORDER', 'PURCHASE_ORDER', 'VENDOR_BILL']);
export const DocumentStatusEnum = z.enum(['DRAFT', 'SIGNED', 'SENT', 'ACCEPTED', 'REJECTED', 'VOIDED']);

export type DocumentType = z.infer<typeof DocumentTypeEnum>;
export type DocumentStatus = z.infer<typeof DocumentStatusEnum>;

// Line Item Schema
export const documentLineSchema = z.object({
    id: z.string().uuid().optional(),
    document_id: z.string().uuid().optional(),
    product_id: z.string().uuid().optional().nullable(),

    description: z.string().min(1, "Descripción requerida"),
    qty: z.number().min(0.0001, "Cantidad debe ser mayor a 0"),
    unit_price: z.number().min(0, "Precio no puede ser negativo"),

    // Calculated fields (frontend validation mostly, backend re-calculates)
    line_total: z.number(),

    // Tax Config stored as JSONB
    // Simple structure for V3: { id: string, name: string, rate: number }[]
    tax_config: z.any().optional(),
});

export type DocumentLine = z.infer<typeof documentLineSchema>;

// Electronic Document Schema (DIAN)
export const electronicDocumentSchema = z.object({
    id: z.string().uuid().optional(),
    document_id: z.string().uuid(),
    environment: z.string(),
    cufe: z.string().optional(),
    xml_url: z.string().optional(),
    qr_data: z.string().optional(),
    dian_status: z.string(),
    rejection_reasons: z.string().optional(),
    sent_at: z.string().optional(),
});

export type ElectronicDocument = z.infer<typeof electronicDocumentSchema>;

// Header Schema
export const documentSchema = z.object({
    id: z.string().uuid().optional(),
    tenant_id: z.string().uuid().optional(),

    doc_type: DocumentTypeEnum,
    number: z.string().optional(), // Generated by backend

    party_id: z.string().uuid("Debes seleccionar un cliente/tercero"),

    issue_date: z.string().refine((val) => !isNaN(Date.parse(val)), "Fecha inválida"),
    due_date: z.string().optional().nullable(),

    currency: z.string().default('COP'),

    subtotal: z.number().default(0),
    taxes: z.number().default(0),
    total: z.number().default(0),

    status: DocumentStatusEnum.default('DRAFT'),

    parent_id: z.string().uuid().optional().nullable(),
    notes_internal: z.string().optional().nullable(),
    notes_public: z.string().optional().nullable(),

    created_at: z.string().optional(),

    // Relations (optional for fetching)
    party: z.custom<Party>().optional(),
    lines: z.array(documentLineSchema).optional(),
    electronic_document: electronicDocumentSchema.optional().nullable(),
});

export type Document = z.infer<typeof documentSchema>;

export const documentFilterSchema = z.object({
    search: z.string().optional(),
    type: DocumentTypeEnum.optional(),
    status: DocumentStatusEnum.optional(),
    start_date: z.string().optional(),
    end_date: z.string().optional(),
    page: z.number().default(1),
    per_page: z.number().default(10),
});

export type DocumentFilters = z.infer<typeof documentFilterSchema>;